Refactor UNLINK
- split off endpoint specific code for better bookkeeping
- minimize aborts for multiple UNLINKs
