+++ Usbipd/RequestReply.cs
@@ -1,0 +1,7 @@
+﻿// SPDX-FileCopyrightText: 2023 Frans van Dorsselaer
+//
+// SPDX-License-Identifier: GPL-3.0-only
+
+namespace Usbipd;
+
+readonly record struct RequestReply(uint seqnum, byte[] bytes);
